WIP - final commit - no more time to spend here
